move languages model to external file
