+++ src/config/languages.ts
@@ -1,0 +1,20 @@
+import polish from "../assets/img/flag-polish.svg";
+import english from "../assets/img/flag-english.svg";
+
+type ConfiguredLanguage = {
+  code: string
+  icon: string
+  name: string
+}
+export const configuredLanguages: ConfiguredLanguage[] = [
+  {
+    code: 'pl',
+    icon: polish,
+    name: 'polski',
+  },
+  {
+    code: 'en',
+    icon: english,
+    name: 'english'
+  }
+]
